+++ src/ui/widget.h
@@ -28,7 +28,7 @@ public slots:
     void updateStackThres(double newThres);
     void resetView();
     void saveImage();
-    void scaleImage(int scaleFactor);
+    void scaleImage(double scaleFactor);
 
 private:
     Ui::Widget *ui;
